+++ app/actions/index.js
@@ -322,6 +322,27 @@ export function fetchRegionTrends(parameter: string, season: string) {
     }
 }
 
+export const COUNT_NUMBER_DATAPOINTS = 'COUNT_NUMBER_DATAPOINTS';
+export function countNumberPoints(countLink: string) {
+    return (dispatch: Dispatch) => {
+        const result = fetch(countLink).then(response => {
+            const json = response.json();
+            return json;
+        })
+            .then(json => {
+                if (json) {
+                    dispatch({
+                        type: COUNT_NUMBER_DATAPOINTS,
+                        number_datapoints: json.datapointsLength,
+                    })
+                }
+            }).catch((error) => {
+                console.log('An ERROR occurred! ' + error);
+            });
+        return result;
+    }
+}
+
 export const ADD_REGION_DETAIL_TRENDS = 'ADD_REGION_DETAIL_TRENDS';
 export function fetchRegionDetailTrends(parameter: string, season: string, region: string) {
     return (dispatch: Dispatch, getState: GetState) => {
@@ -433,6 +454,55 @@ export function addSearchLocation(location: ? string) {
             selected_filters
         });
         const idx = selected_filters.indexOf('locations');
+        dispatch(updateAvailableSensors(idx));
+    }
+}
+
+export const ADD_SEARCH_ONLINE = 'ADD_SEARCH_ONLINE';
+export function addSearchOnline(online: ? string) {
+    return (dispatch: Dispatch, getState: GetState) => {
+        const state = getState();
+        const selected_filters = state.searchFilters.selected;
+        dispatch({
+            type: ADD_SEARCH_ONLINE,
+            online,
+            selected_filters
+        });
+        const idx = selected_filters.indexOf('online');
+        dispatch(updateAvailableSensors(idx));
+    }
+}
+
+export const ADD_SPAN_START = 'ADD_SPAN_START';
+export function addSpanStart(span: ?Date) {
+    return (dispatch: Dispatch, getState: GetState) => {
+        const state = getState();
+        const selected_filters = state.searchFilters.selected;
+        const availableSensors = state.sensors.available_sensors;
+        dispatch({
+            type: ADD_SPAN_START,
+            span,
+            selected_filters,
+            availableSensors
+        });
+        const idx = selected_filters.indexOf('span');
+        dispatch(updateAvailableSensors(idx));
+    }
+}
+
+export const ADD_SPAN_END = 'ADD_SPAN_END';
+export function addSpanEnd(span: ?Date) {
+    return (dispatch: Dispatch, getState: GetState) => {
+        const state = getState();
+        const selected_filters = state.searchFilters.selected;
+        const availableSensors = state.sensors.available_sensors;
+        dispatch({
+            type: ADD_SPAN_END,
+            span,
+            selected_filters,
+            availableSensors
+        });
+        const idx = selected_filters.indexOf('span');
         dispatch(updateAvailableSensors(idx));
     }
 }
@@ -551,7 +621,6 @@ export function updateAvailableSensors(idx: number) {
         })
     }
 }
-
 
 export const RECEIVE_PARAMETERS = "RECEIVE_PARAMETERS";
 export const RECEIVE_MULTI_PARAMETERS = "RECEIVE_MULTI_PARAMETERS";
@@ -1013,3 +1082,4 @@ export function analysisSavedSearch(
         })
     }
 }
+