+++ app/containers/DownloadButtons.js
@@ -4,6 +4,8 @@
 
 import {connect} from 'react-redux';
 import DownloadButtonsComponent from '../components/DownloadButtons';
+import {countNumberPoints} from "../actions";
+import type {Dispatch} from "../utils/flowtype";
 
 const mapStateToProps = (state) => {
     return {
@@ -16,10 +18,20 @@ const mapStateToProps = (state) => {
         api: state.backends.selected,
         drawShapeCoordinates: state.sensors.shape_coordinates,
         showSensors: state.sensors.search_sensors,
-        multi_parameter_map: state.parameters.multi_parameter_map
+        availableSensors: state.sensors.available_sensors,
+        multi_parameter_map: state.parameters.multi_parameter_map,
+        numberPoints: state.sensors.number_datapoints
     }
 };
 
-const DownloadButtons = connect(mapStateToProps)(DownloadButtonsComponent);
+const mapDispatchToProps = (dispatch: Dispatch) => {
+    return {
+        onSelectDownload: (countLink) => {
+            dispatch(countNumberPoints(countLink));
+        }
+    }
+};
+
+const DownloadButtons = connect(mapStateToProps, mapDispatchToProps)(DownloadButtonsComponent);
 
 export default DownloadButtons;