+++ app/reducers/sensors.js
@@ -4,7 +4,7 @@
 
 import {
     RECEIVE_SENSORS, UPDATE_AVAILABLE_SENSORS, UPDATE_EXPLORE_SENSORS, RESET_EXPLORE_SENSORS,
-    ADD_CUSTOM_LOCATION_FILTER, CLEAR_SENSORS, RECEIVE_MULTI_PARAMETERS
+    ADD_CUSTOM_LOCATION_FILTER, CLEAR_SENSORS, RECEIVE_MULTI_PARAMETERS, COUNT_NUMBER_DATAPOINTS
 } from '../actions';
 import type {Sensor, Sensors, sensorsState, MapWithLabel, MapWithLabels, Parameters} from '../utils/flowtype';
 import {
@@ -22,6 +22,7 @@ type SensorAction = {|
     multi_parameter_map: { [string]: Array<string> },
     shape_coordinates: Array<number>,
     explore_filtering: Object,
+    number_datapoints: number
 |};
 
 const defaultState = {
@@ -61,6 +62,11 @@ const sensors = (state: sensorsState = defaultState, action: SensorAction) => {
             const copy_sensors = state.sensors.slice(0);
             return Object.assign({}, state, {
                 parameters: collectParameters(copy_sensors, action.parameters, action.multi_parameter_map)
+            });
+
+        case COUNT_NUMBER_DATAPOINTS:
+            return Object.assign({}, state, {
+                number_datapoints: action.number_datapoints
             });
 
         case UPDATE_AVAILABLE_SENSORS:
@@ -256,6 +262,43 @@ export function collectDates(sensorsData: Sensors): CollectDate {
     return {'start': minDate, 'end': maxDate};
 }
 
+export function collectSpanDates(sensorsData: Sensors): CollectDate {
+    let minDate = new Date();
+    let maxDate = new Date("1983-01-01");
+
+    sensorsData.map(s => {
+        const sensorStartTime = new Date(s.min_start_time);
+        const sensorEndTime = new Date(s.max_end_time);
+        if (sensorStartTime.getTime() < minDate.getTime()) {
+            minDate = sensorStartTime;
+        }
+        if (sensorEndTime.getTime() > maxDate) {
+            maxDate = sensorEndTime;
+        }
+    });
+
+    return {'start': minDate, 'end': maxDate};
+}
+
+export function collectOnlineOffline(sensorsData: Sensors): MapWithLabels {
+    let status = [];
+    let onlineAvailable = false;
+
+    sensorsData.map(s => {
+        if (s.properties.online_status !== undefined && onlineAvailable === false) {
+            onlineAvailable = true;
+        }
+    });
+    if (onlineAvailable === true) {
+        status.push({'id': "online", 'label': "Online"});
+        status.push({'id': "offline", 'label': "Offline"});
+        status.push({'id': "none", 'label': "All Options"});
+    } else {
+        status.push({'id': "none", 'label': "All Options"});
+    }
+    return status;
+}
+
 export function collectLocations(sensorsData: Sensors): MapWithLabels {
     let locations = [];
     const additional_location = window.configruntime.gd3.additional_locations;
@@ -397,6 +440,40 @@ function filterAvailableSensors(state: sensorsState, selectedFilters: Array<stri
                 }
 
                 return;
+
+            case 'span':
+                if (selectedSearch.span.selected.start !== "" && selectedSearch.span.selected.end !== ""
+                    && selectedSearch.span.selected.start !== null && selectedSearch.span.selected.end !== null) {
+                    new_sensors = [];
+                    av_sensors.map((sensor) => {
+                        if (selectedSearch.span.selected.end <= new Date(sensor.max_end_time) &&
+                            selectedSearch.span.selected.start >= new Date(sensor.min_start_time)) {
+                            new_sensors.push(sensor);
+                        }
+                    });
+                    av_sensors = new_sensors;
+                }
+                return;
+
+            case 'online':
+                // This uses the radio button's value
+                if (selectedSearch.online.selected !== null) {
+                    new_sensors = [];
+                    if (selectedSearch.online.selected === "eitherOption" || selectedSearch.online.selected === "none") {
+                        new_sensors = av_sensors;
+                    } else {
+                        av_sensors.map((sensor => {
+                            if (sensor.properties.online_status !== undefined) {
+                                if (sensor.properties.online_status === selectedSearch.online.selected) {
+                                    new_sensors.push(sensor);
+                                }
+                            }
+                        }));
+                    }
+                    av_sensors = new_sensors;
+                }
+                return;
+
         }
 
     });