+++ app/utils/getConfig.js
@@ -47,6 +47,15 @@ export function getCustomLocation(location: string): Object {
         });
 }
 
+export function getAreaLocations(location: string): Object {
+    const sensors_regions = window.configruntime.gd3.sensors_regions;
+
+    return sensors_regions.find(
+        function (custom_location) {
+            return custom_location.properties.id === location;
+        });
+}
+
 export function getLakesOrdering(key: string) {
     const additional_locations = window.configruntime.gd3.additional_locations;
     let order = {};
@@ -465,6 +474,38 @@ export function getDownloadButtonPath() {
     return getValue;
 }
 
+export function getDownloadButtonPathCount() {
+    let getValue = '/geostreams/api/datapoints?';
+    if (window.configruntime.gd3.download_button_path_count) {
+        getValue = window.configruntime.gd3.download_button_path_count;
+    }
+    return getValue;
+}
+
+export function getDownloadMaxDatapointsAllowed() {
+    let getValue = 500;
+    if (window.configruntime.gd3.download_count_max_datapoints) {
+        getValue = window.configruntime.gd3.download_count_max_datapoints;
+    }
+    return getValue;
+}
+
+export function getGeneralDownloadErrorText() {
+    let getValue = 'An ERROR occurred with Download - Please try again!';
+    if (window.configruntime.gd3.general_error) {
+        getValue = window.configruntime.gd3.general_error;
+    }
+    return getValue;
+}
+
+export function getDatapointsDownloadErrorText() {
+    let getValue = 'Too Many Datapoints to Download - Please try again!';
+    if (window.configruntime.gd3.max_datapoints_error) {
+        getValue = window.configruntime.gd3.max_datapoints_error;
+    }
+    return getValue;
+}
+
 export function getTimeSeriesZeroStart() {
     return window.configruntime.gd3.time_series_zero_start;
 }
@@ -576,6 +617,14 @@ export function getCarouselImageNames() {
     return nameList;
 }
 
+export function getCarouselImageCaptions() {
+    let captionList = '';
+    if (window.configruntime.gd3.home_page_carousel_captions) {
+        captionList = window.configruntime.gd3.home_page_carousel_captions;
+    }
+    return captionList;
+}
+
 export function getMapPopupZoom() {
     let getValue = 10;
     if (window.configruntime.gd3.mapPopupZoomMax) {
@@ -588,6 +637,14 @@ export function getLoadingTimeLimit() {
     let getValue = 100000;
     if (window.configruntime.gd3.load_time_limit) {
         getValue = window.configruntime.gd3.load_time_limit;
+    }
+    return getValue;
+}
+
+export function getIntervalTime() {
+    let getValue = 3000;
+    if (window.configruntime.gd3.set_interval_time) {
+        getValue = window.configruntime.gd3.set_interval_time;
     }
     return getValue;
 }
@@ -634,7 +691,7 @@ export function useAnalysisSearches() {
     return getValue;
 }
 
-export function getAnalysisSearchHeading(){
+export function getAnalysisSearchHeading() {
     let getValue = 'Saved Searches: ';
     if (window.configruntime.gd3.analysis_search_heading) {
         getValue = window.configruntime.gd3.analysis_search_heading;